remove EventSpan
